fix(alerts): split bounded projection helpers
